refactor(tests): split into build-driver + test-app to remove mopro-ffi from host build
tests/ is now a pure build-pipeline driver (no lib code, no mopro-ffi dep).
test-app/ holds the actual library that gets compiled for iOS/Android/Wasm.
build_from_env_at(project_dir) added to mopro-build-common; each backend
exposes build_at(project_dir) so tests/build.rs can point at test-app/.

This avoids compiling mopro-ffi and linking cdylib/staticlib for the host
when running `cargo build -p tests`, significantly reducing CI cold-build time.
