+++ ffi/uniffi/src/android.rs
@@ -21,6 +21,10 @@ pub fn build() {
     build_from_env::<AndroidPlatform>()
 }
 
+pub fn build_at(project_dir: &std::path::Path) {
+    mopro_build_common::build_from_env_at::<AndroidPlatform>(project_dir)
+}
+
 #[derive(Default, Clone)]
 pub struct ArchBuildConfig {
     pub extra_env: Vec<(String, String)>,
