+++ tests/build.rs
@@ -1,8 +1,8 @@
 // Build-pipeline driver for the mopro test project.
 //
 // Each backend only runs if the corresponding arch env var is set.
-// This lets `cargo build` succeed on any machine without toolchain requirements,
-// while Xcode / CI / wasm-pack workflows opt in via env vars.
+// The library being compiled lives in test-app/ (which carries the mopro-ffi
+// dependency); this crate only orchestrates the build pipeline.
 //
 // iOS:          IOS_ARCHS=aarch64-apple-ios,aarch64-apple-ios-sim  cargo build
 // Android:      ANDROID_ARCHS=aarch64-linux-android                cargo build
@@ -11,19 +11,22 @@
 // React Native: REACT_NATIVE_ARCHS=aarch64-apple-ios              cargo build
 
 fn main() {
+    let manifest_dir = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+    let app_dir = manifest_dir.parent().unwrap().join("test-app");
+
     if std::env::var("IOS_ARCHS").is_ok() {
-        mopro_uniffi_backend::ios::build();
+        mopro_uniffi_backend::ios::build_at(&app_dir);
     }
     if std::env::var("ANDROID_ARCHS").is_ok() {
-        mopro_uniffi_backend::android::build();
+        mopro_uniffi_backend::android::build_at(&app_dir);
     }
     if std::env::var("WEB_ARCHS").is_ok() {
-        mopro_wasm_backend::build();
+        mopro_wasm_backend::build_at(&app_dir);
     }
     if std::env::var("FLUTTER_ARCHS").is_ok() {
-        mopro_flutter_backend::build();
+        mopro_flutter_backend::build_at(&app_dir);
     }
     if std::env::var("REACT_NATIVE_ARCHS").is_ok() {
-        mopro_react_native_backend::build();
+        mopro_react_native_backend::build_at(&app_dir);
     }
 }
